ffmpeg_muxer bugfix

+++ services/media_engine/plugins/ffmpeg_adapter/muxer/ffmpeg_muxer_plugin.cpp
@@ -733,7 +733,7 @@ Status FFmpegMuxerPlugin::AddTimedMetaTrack(
 {
     int32_t srcTrackID = -1;
     bool ret = trackDesc->Get<Tag::TIMED_METADATA_SRC_TRACK>(srcTrackID); // source track id
-    FALSE_RETURN_V_MSG_E(ret && srcTrackID >= 0 && srcTrackID < formatContext_->nb_streams,
+    FALSE_RETURN_V_MSG_E(ret && srcTrackID >= 0 && static_cast<uint32_t>(srcTrackID) < formatContext_->nb_streams,
         Status::ERROR_MISMATCHED_TYPE, "timed_metadata_track_id %{public}d is invalid", srcTrackID);
 
     auto st = avformat_new_stream(formatContext_.get(), nullptr);
